pyproject.toml
[tool.poetry]
name = "vidvlog"
version = "0.1.0"
description = ""
authors = ["reinanbr <perseu912@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
moviepy = "^1.0.3"
bing-image-downloader = "^1.1.2"
opencv-python = "^4.10.0.84"
pylint = "^3.3.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
jedi = "^0.19.2"
python-dotenv = "^1.0.1"
kitano = "^0.0.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
